Rename NetSimGlobals.js => netsimGlobals.js
