+++ apps/src/netsim/netsimGlobals.js
